# task/swagger.yaml
swagger: "2.0"
info:
  description: "GitHub Release Notification API that allows users to subscribe to email notifications about new releases of a chosen GitHub repository."
  version: "1.0.0"
  title: "GitHub Release Notification API"
host: "127.0.0.1:3000"
...
basePath: "/api"
...
schemes:
  - "http"
  - "https"
paths:
  /subscribe:
    post:
      tags:
        - "subscription"
      summary: "Subscribe to release notifications"
      description: "Subscribe an email to receive notifications about new releases of a GitHub repository. The repository is validated via GitHub API."
      operationId: "subscribe"
      consumes:
        - "application/json"
        - "application/x-www-form-urlencoded"
      produces:
        - "application/json"
      parameters:
        - name: "email"
          in: "formData"
          description: "Email address to subscribe"
          required: true
          type: "string"
        - name: "repo"
          in: "formData"
          description: "GitHub repository in owner/repo format (e.g., golang/go)"
          required: true
          type: "string"
      responses:
        "200":
          description: "Subscription successful. Confirmation email sent."
        "400":
          description: "Invalid input (e.g., invalid repo format)"
        "404":
          description: "Repository not found on GitHub"
        "409":
          description: "Email already subscribed to this repository"
...